+++ crates/stackable-webhook/src/servers/conversion.rs
@@ -1,32 +1,21 @@
 use std::{fmt::Debug, net::SocketAddr};
 
 use axum::{Json, Router, routing::post};
-use k8s_openapi::{
-    ByteString,
-    apiextensions_apiserver::pkg::apis::apiextensions::v1::{
-        CustomResourceConversion, CustomResourceDefinition, ServiceReference, WebhookClientConfig,
-        WebhookConversion,
-    },
-};
+use k8s_openapi::apiextensions_apiserver::pkg::apis::apiextensions::v1::CustomResourceDefinition;
 // Re-export this type because users of the conversion webhook server require
 // this type to write the handler function. Instead of importing this type from
 // kube directly, consumers can use this type instead. This also eliminates
 // keeping the kube dependency version in sync between here and the operator.
 pub use kube::core::conversion::ConversionReview;
-use kube::{
-    Api, Client, ResourceExt,
-    api::{Patch, PatchParams},
-};
-use snafu::{OptionExt, ResultExt, Snafu};
-use tokio::{sync::mpsc, try_join};
+use kube::{Client, ResourceExt};
+use snafu::{ResultExt, Snafu};
+use tokio::sync::{mpsc, oneshot};
 use tracing::instrument;
-use x509_cert::{
-    Certificate,
-    der::{EncodePem, pem::LineEnding},
-};
+use x509_cert::Certificate;
 
 use crate::{
-    WebhookError, WebhookHandler, WebhookServer, constants::CONVERSION_WEBHOOK_HTTPS_PORT,
+    WebhookError, WebhookHandler, WebhookServer,
+    maintainer::{CustomResourceDefinitionMaintainer, CustomResourceDefinitionMaintainerOptions},
     options::WebhookOptions,
 };
 
@@ -68,126 +57,98 @@ where
 
 // TODO: Add a builder, maybe with `bon`.
 #[derive(Debug)]
-pub struct ConversionWebhookOptions {
+pub struct ConversionWebhookOptions<'a> {
     /// The bind address to bind the HTTPS server to.
     pub socket_addr: SocketAddr,
 
     /// The namespace the operator/webhook is running in.
-    pub namespace: String,
+    pub namespace: &'a str,
 
     /// The name of the Kubernetes service which points to the operator/webhook.
-    pub service_name: String,
-
-    /// If the CRDs should be maintained automatically. Use the (negated) value from
-    /// `stackable_operator::cli::ProductOperatorRun::disable_crd_maintenance`
-    /// for this.
-    // # Because of https://github.com/rust-lang/cargo/issues/3475 we can not use a real link here
-    pub maintain_crds: bool,
-
-    /// The field manager used to apply Kubernetes objects, typically the operator name, e.g.
-    /// `airflow-operator`.
-    pub field_manager: String,
+    pub service_name: &'a str,
 }
 
 /// A ready-to-use CRD conversion webhook server.
 ///
 /// See [`ConversionWebhookServer::new()`] for usage examples.
-pub struct ConversionWebhookServer {
-    crds: Vec<CustomResourceDefinition>,
-    options: ConversionWebhookOptions,
-    router: Router,
-    client: Client,
-}
+pub struct ConversionWebhookServer(WebhookServer);
 
 impl ConversionWebhookServer {
-    /// Creates a new conversion webhook server, which expects POST requests being made to the
-    /// `/convert/{crd name}` endpoint.
+    /// The default socket address the conversion webhook server binds to, see
+    /// [`WebhookServer::DEFAULT_SOCKET_ADDRESS`].
+    pub const DEFAULT_SOCKET_ADDRESS: SocketAddr = WebhookServer::DEFAULT_SOCKET_ADDRESS;
+
+    /// Creates and returns a new [`ConversionWebhookServer`], which expects POST requests being
+    /// made to the `/convert/{CRD_NAME}` endpoint.
     ///
-    /// You need to provide a few things for every CRD passed in via the `crds_and_handlers` argument:
+    /// The TLS certificate is automatically generated and rotated.
     ///
-    /// 1. The CRD
-    /// 2. A conversion function to convert between CRD versions. Typically you would use the
-    ///    the auto-generated `try_convert` function on CRD spec definition structs for this.
-    /// 3. A [`kube::Client`] used to create/update the CRDs.
+    /// ## Parameters
     ///
-    /// The [`ConversionWebhookServer`] takes care of reconciling the CRDs into the Kubernetes
-    /// cluster and takes care of adding itself as conversion webhook. This includes TLS
-    /// certificates and CA bundles.
+    /// This function expects the following parameters:
     ///
-    /// # Example
+    /// - `crds_and_handlers`: An iterator over a 2-tuple (pair) mapping a [`CustomResourceDefinition`]
+    ///   to a handler function. In most cases, the generated `CustomResource::try_merge` function
+    ///   should be used. It provides the expected `fn(ConversionReview) -> ConversionReview`
+    ///   signature.
+    /// - `options`: Provides [`ConversionWebhookOptions`] to customize various parts of the
+    ///   webhook server, eg. the socket address used to listen on.
+    ///
+    /// ## Return Values
+    ///
+    /// This function returns a [`Result`] which contains a 2-tuple (pair) of values for the [`Ok`]
+    /// variant:
+    ///
+    /// - The [`ConversionWebhookServer`] itself. This is used to run the server. See
+    ///   [`ConversionWebhookServer::run`] for more details.
+    /// - The [`mpsc::Receiver`] which will be used to send out messages containing the newly
+    ///   generated TLS certificate. This channel is used by the CRD maintainer to trigger a
+    ///   reconcile of the CRDs it maintains.
+    ///
+    /// ## Example
     ///
     /// ```no_run
-    /// use clap::Parser;
-    /// use stackable_webhook::{
-    ///     servers::{ConversionWebhookServer, ConversionWebhookOptions},
-    ///     constants::CONVERSION_WEBHOOK_HTTPS_PORT,
-    ///     WebhookOptions
-    /// };
-    /// use stackable_operator::{
-    ///     kube::Client,
-    ///     crd::s3::{S3Connection, S3ConnectionVersion},
-    ///     cli::{RunArguments, MaintenanceOptions},
-    /// };
+    /// # use tokio_rustls::rustls::crypto::{CryptoProvider, ring::default_provider};
+    /// use stackable_webhook::servers::{ConversionWebhookServer, ConversionWebhookOptions};
+    /// use stackable_operator::crd::s3::{S3Connection, S3ConnectionVersion};
     ///
-    /// # async fn test() {
-    /// // Things that should already be in you operator:
-    /// const OPERATOR_NAME: &str = "product-operator";
-    /// let client = Client::try_default().await.expect("failed to create Kubernetes client");
-    /// let RunArguments {
-    ///     operator_environment,
-    ///     maintenance: MaintenanceOptions {
-    ///         disable_crd_maintenance,
-    ///         ..
-    ///     },
-    ///     ..
-    /// } = RunArguments::parse();
-    ///
-    ///  let crds_and_handlers = [
+    /// # #[tokio::main]
+    /// # async fn main() {
+    /// # CryptoProvider::install_default(default_provider()).unwrap();
+    /// let crds_and_handlers = vec![
     ///     (
     ///         S3Connection::merged_crd(S3ConnectionVersion::V1Alpha1)
-    ///             .expect("failed to merge S3Connection CRD"),
-    ///         S3Connection::try_convert as fn(_) -> _,
-    ///     ),
+    ///             .expect("the S3Connection CRD must be merged"),
+    ///         S3Connection::try_convert,
+    ///     )
     /// ];
     ///
     /// let options = ConversionWebhookOptions {
-    ///     socket_addr: format!("0.0.0.0:{CONVERSION_WEBHOOK_HTTPS_PORT}")
-    ///         .parse()
-    ///         .expect("static address is always valid"),
-    ///     namespace: operator_environment.operator_namespace,
-    ///     service_name: operator_environment.operator_service_name,
-    ///     maintain_crds: !disable_crd_maintenance,
-    ///     field_manager: OPERATOR_NAME.to_owned(),
+    ///     socket_addr: ConversionWebhookServer::DEFAULT_SOCKET_ADDRESS,
+    ///     namespace: "stackable-operators",
+    ///     service_name: "product-operator",
     /// };
     ///
-    /// // Construct the conversion webhook server
-    /// let conversion_webhook = ConversionWebhookServer::new(
-    ///     crds_and_handlers,
-    ///     options,
-    ///     client,
-    /// )
-    /// .await
-    /// .expect("failed to create ConversionWebhookServer");
+    /// let (conversion_webhook_server, _certificate_rx) =
+    ///         ConversionWebhookServer::new(crds_and_handlers, options)
+    ///             .await
+    ///             .unwrap();
     ///
-    /// conversion_webhook.run().await.expect("failed to run ConversionWebhookServer");
+    /// conversion_webhook_server.run().await.unwrap();
     /// # }
     /// ```
-    #[instrument(
-        name = "create_conversion_webhook_server",
-        skip(crds_and_handlers, client)
-    )]
+    #[instrument(name = "create_conversion_webhook_server", skip(crds_and_handlers))]
     pub async fn new<H>(
         crds_and_handlers: impl IntoIterator<Item = (CustomResourceDefinition, H)>,
-        options: ConversionWebhookOptions,
-        client: Client,
-    ) -> Result<Self, ConversionWebhookError>
+        options: ConversionWebhookOptions<'_>,
+    ) -> Result<(Self, mpsc::Receiver<Certificate>), ConversionWebhookError>
     where
         H: WebhookHandler<ConversionReview, ConversionReview> + Clone + Send + Sync + 'static,
     {
         tracing::debug!("create new conversion webhook server");
 
         let mut router = Router::new();
-        let mut crds = Vec::new();
+
         for (crd, handler) in crds_and_handlers {
             let crd_name = crd.name_any();
             let handler_fn = |Json(review): Json<ConversionReview>| async {
@@ -195,39 +156,21 @@ impl ConversionWebhookServer {
                 Json(review)
             };
 
+            // TODO (@Techassi): Make this part of the trait mentioned above
             let route = format!("/convert/{crd_name}");
             router = router.route(&route, post(handler_fn));
-            crds.push(crd);
         }
-
-        Ok(Self {
-            options,
-            router,
-            client,
-            crds,
-        })
-    }
-
-    pub async fn run(self) -> Result<(), ConversionWebhookError> {
-        tracing::info!("starting conversion webhook server");
-
-        let Self {
-            options,
-            router,
-            client,
-            crds,
-        } = self;
 
         let ConversionWebhookOptions {
             socket_addr,
             namespace: operator_namespace,
             service_name: operator_service_name,
-            maintain_crds,
-            field_manager,
         } = &options;
 
         // This is how Kubernetes calls us, so it decides about the naming.
         // AFAIK we can not influence this, so this is the only SAN entry needed.
+        // TODO (@Techassi): The cluster domain should be included here, so that (non Kubernetes)
+        // HTTP clients can use the FQDN of the service for testing or user use-cases.
         let subject_alterative_dns_name =
             format!("{operator_service_name}.{operator_namespace}.svc",);
 
@@ -236,127 +179,144 @@ impl ConversionWebhookServer {
             socket_addr: *socket_addr,
         };
 
-        let (server, mut cert_rx) = WebhookServer::new(router, webhook_options)
+        let (server, certificate_rx) = WebhookServer::new(router, webhook_options)
             .await
             .context(CreateWebhookServerSnafu)?;
 
-        // We block the ConversionWebhookServer creation until the certificates have been generated.
-        // This way we
-        // 1. Are able to apply the CRDs before we start the actual controllers relying on them
-        // 2. Avoid updating them shortly after as cert have been generated. Doing so would cause
-        // unnecessary "too old resource version" errors in the controllers as the CRD was updated.
-        let current_cert = cert_rx
-            .recv()
-            .await
-            .context(ReceiveCertificateFromChannelSnafu)?;
+        Ok((Self(server), certificate_rx))
+    }
 
-        if *maintain_crds {
-            Self::reconcile_crds(
-                &client,
-                field_manager,
-                &crds,
-                operator_namespace,
-                operator_service_name,
-                current_cert,
-            )
-            .await
-            .context(ReconcileCrdsSnafu)?;
+    /// Creates and returns a tuple consisting of a [`ConversionWebhookServer`], a [`CustomResourceDefinitionMaintainer`],
+    /// and a [`oneshot::Receiver`].
+    ///
+    /// ## Parameters
+    ///
+    /// - `crds_and_handlers`: An iterator over a 2-tuple (pair) mapping a [`CustomResourceDefinition`]
+    ///   to a handler function. In most cases, the generated `CustomResource::try_merge` function
+    ///   should be used. It provides the expected `fn(ConversionReview) -> ConversionReview`
+    ///   signature.
+    /// - `operator_service_name`: The name of the Kubernetes service name which points to the
+    ///   operator/conversion webhook. This is used to construct the service reference in the CRD
+    ///   `spec.conversion` field.
+    /// - `operator_namespace`: The namespace the operator runs in. This is used to construct the
+    ///   service reference in the CRD `spec.conversion` field.
+    /// - `disable_maintainer`: A boolean value to indicate if the [`CustomResourceDefinitionMaintainer`]
+    ///   should be disabled.
+    /// - `client`: A [`kube::Client`] used to maintain the custom resource definitions.
+    ///
+    /// See the referenced items for more details on usage.
+    ///
+    /// ## Return Values
+    ///
+    /// - The [`ConversionWebhookServer`] itself. This is used to run the server. See
+    ///   [`ConversionWebhookServer::run`] for more details.
+    /// - The [`CustomResourceDefinitionMaintainer`] which is used to run the maintainer. See
+    ///   [`CustomResourceDefinitionMaintainer::run`] for more details.
+    /// - A [`oneshot::Receiver`] which is triggered after the initial reconciliation of the CRDs
+    ///   succeeded. This signal can be used to deploy any custom resources defined by these CRDs.
+    ///
+    /// ## Example
+    ///
+    /// ```no_run
+    /// # use futures_util::TryFutureExt;
+    /// # use tokio_rustls::rustls::crypto::{CryptoProvider, ring::default_provider};
+    /// use stackable_webhook::servers::{ConversionWebhookServer, ConversionWebhookOptions};
+    /// use stackable_operator::{kube::Client, crd::s3::{S3Connection, S3ConnectionVersion}};
+    ///
+    /// # #[tokio::main]
+    /// # async fn main() {
+    /// # CryptoProvider::install_default(default_provider()).unwrap();
+    /// let client = Client::try_default().await.unwrap();
+    ///
+    /// let crds_and_handlers = vec![
+    ///     (
+    ///         S3Connection::merged_crd(S3ConnectionVersion::V1Alpha1)
+    ///             .expect("the S3Connection CRD must be merged"),
+    ///         S3Connection::try_convert,
+    ///     )
+    /// ];
+    ///
+    /// let (conversion_webhook_server, crd_maintainer, _initial_reconcile_rx) =
+    ///     ConversionWebhookServer::with_maintainer(
+    ///         crds_and_handlers,
+    ///         "my-operator",
+    ///         "my-namespace",
+    ///         "my-field-manager",
+    ///         false,
+    ///         client,
+    ///     )
+    ///     .await
+    ///     .unwrap();
+    ///
+    /// let conversion_webhook_server = conversion_webhook_server
+    ///     .run()
+    ///     .map_err(|err| err.to_string());
+    ///
+    /// let crd_maintainer = crd_maintainer
+    ///     .run()
+    ///     .map_err(|err| err.to_string());
+    ///
+    /// // Run both the conversion webhook server and crd_maintainer concurrently, eg. with
+    /// // futures::try_join!.
+    /// futures_util::try_join!(conversion_webhook_server, crd_maintainer).unwrap();
+    /// # }
+    /// ```
+    pub async fn with_maintainer<'a, H>(
+        // TODO (@Techassi): Use a trait type here which can be used to build all part of the
+        // conversion webhook server and a CRD maintainer.
+        crds_and_handlers: impl IntoIterator<Item = (CustomResourceDefinition, H)> + Clone,
+        operator_service_name: &'a str,
+        operator_namespace: &'a str,
+        field_manager: &'a str,
+        disable_maintainer: bool,
+        client: Client,
+    ) -> Result<
+        (
+            Self,
+            CustomResourceDefinitionMaintainer<'a>,
+            oneshot::Receiver<()>,
+        ),
+        ConversionWebhookError,
+    >
+    where
+        H: WebhookHandler<ConversionReview, ConversionReview> + Clone + Send + Sync + 'static,
+    {
+        let socket_addr = ConversionWebhookServer::DEFAULT_SOCKET_ADDRESS;
 
-            try_join!(
-                Self::run_webhook_server(server),
-                Self::run_crd_reconciliation_loop(
-                    cert_rx,
-                    &client,
-                    field_manager,
-                    &crds,
-                    operator_namespace,
-                    operator_service_name,
-                ),
-            )?;
-        } else {
-            Self::run_webhook_server(server).await?;
+        // TODO (@Techassi): These should be moved into a builder
+        let webhook_options = ConversionWebhookOptions {
+            service_name: operator_service_name,
+            namespace: operator_namespace,
+            socket_addr,
         };
 
-        Ok(())
-    }
+        let (conversion_webhook_server, certificate_rx) =
+            Self::new(crds_and_handlers.clone(), webhook_options).await?;
 
-    async fn run_webhook_server(server: WebhookServer) -> Result<(), ConversionWebhookError> {
-        server.run().await.context(RunWebhookServerSnafu)
-    }
+        let definitions = crds_and_handlers.into_iter().map(|(crd, _)| crd);
 
-    async fn run_crd_reconciliation_loop(
-        mut cert_rx: mpsc::Receiver<Certificate>,
-        client: &Client,
-        field_manager: &str,
-        crds: &[CustomResourceDefinition],
-        operator_namespace: &str,
-        operator_service_name: &str,
-    ) -> Result<(), ConversionWebhookError> {
-        while let Some(current_cert) = cert_rx.recv().await {
-            Self::reconcile_crds(
-                client,
-                field_manager,
-                crds,
-                operator_namespace,
-                operator_service_name,
-                current_cert,
-            )
-            .await
-            .context(ReconcileCrdsSnafu)?;
-        }
-        Ok(())
-    }
+        // TODO (@Techassi): These should be moved into a builder
+        let maintainer_options = CustomResourceDefinitionMaintainerOptions {
+            webhook_https_port: socket_addr.port(),
+            disabled: disable_maintainer,
+            operator_service_name,
+            operator_namespace,
+            field_manager,
+        };
 
-    #[instrument(skip_all)]
-    async fn reconcile_crds(
-        client: &Client,
-        field_manager: &str,
-        crds: &[CustomResourceDefinition],
-        operator_namespace: &str,
-        operator_service_name: &str,
-        current_cert: Certificate,
-    ) -> Result<(), ConversionWebhookError> {
-        tracing::info!(
-            crds = ?crds.iter().map(CustomResourceDefinition::name_any).collect::<Vec<_>>(),
-            "Reconciling CRDs"
+        let (maintainer, initial_reconcile_rx) = CustomResourceDefinitionMaintainer::new(
+            client,
+            certificate_rx,
+            definitions,
+            maintainer_options,
         );
-        let ca_bundle = current_cert
-            .to_pem(LineEnding::LF)
-            .context(ConvertCaToPemSnafu)?;
 
-        let crd_api: Api<CustomResourceDefinition> = Api::all(client.clone());
-        for mut crd in crds.iter().cloned() {
-            let crd_name = crd.name_any();
+        Ok((conversion_webhook_server, maintainer, initial_reconcile_rx))
+    }
 
-            crd.spec.conversion = Some(CustomResourceConversion {
-                strategy: "Webhook".to_string(),
-                webhook: Some(WebhookConversion {
-                    // conversionReviewVersions indicates what ConversionReview versions are understood/preferred by the webhook.
-                    // The first version in the list understood by the API server is sent to the webhook.
-                    // The webhook must respond with a ConversionReview object in the same version it received.
-                    conversion_review_versions: vec!["v1".to_string()],
-                    client_config: Some(WebhookClientConfig {
-                        service: Some(ServiceReference {
-                            name: operator_service_name.to_owned(),
-                            namespace: operator_namespace.to_owned(),
-                            path: Some(format!("/convert/{crd_name}")),
-                            port: Some(CONVERSION_WEBHOOK_HTTPS_PORT.into()),
-                        }),
-                        ca_bundle: Some(ByteString(ca_bundle.as_bytes().to_vec())),
-                        url: None,
-                    }),
-                }),
-            });
-
-            let patch = Patch::Apply(&crd);
-            let patch_params = PatchParams::apply(field_manager);
-            crd_api
-                .patch(&crd_name, &patch_params, &patch)
-                .await
-                .with_context(|_| UpdateCrdSnafu {
-                    crd_name: crd_name.to_string(),
-                })?;
-        }
-        Ok(())
+    /// Runs the [`ConversionWebhookServer`] asynchronously.
+    pub async fn run(self) -> Result<(), ConversionWebhookError> {
+        tracing::info!("run conversion webhook server");
+        self.0.run().await.context(RunWebhookServerSnafu)
     }
 }